Update README.md
RESTful API development refers to creating an API (Application Programming Interface) that adheres to the principles of REST (Representational State Transfer). REST is an architectural style for designing networked applications, and when developing a RESTful API, the focus is on providing a lightweight, scalable way to interact with data over the internet.